// LinkERP/ClientApp/src/app/models/pur/lbs-pur-landed-cost-shipment-lines.ts
import { BaseEntity } from "../base-entity"
export class LBSPURLandedCostShipmentLines extends BaseEntity {
  CompanyID: any;
  PurchaseLandedCostID: any;
  InvoicesNumber: any;
  PurchaseOrderID: any;
  PurchaseOrderLineID: any;
  VendorID: any;
  WarehouseID:any;
  ProductType: any;
  ProductID: any;
  ProductCode: any;
  Description: any;
  UseSerialNo: any;
  UseExpiry: any;
  Weight: any;
  Volume: any;
  QuantityOrdered: any;
  QuantityReceivedThisShipment: any;
  LandedUnitCostTaxExclusiveHome: any;
  ProductName
  ShipmentNumber
  ShipmentLineID
  ShipmentBookingID
  LineTotalForeignExchangeCostTaxExclusive
  LineTotalTaxExclusiveHome: any;
  LineTotalLandedCostTaxExclusiveHome: any;
  CreatedBy:any;
  //ClassificationID: any;
  // LineTotalTaxInclusiveHome: any;
  // LineTotalTaxExclusiveForeign: any;
  // LineTotalTaxInclusiveForeign: any;
  // ImportCosts: any;
  // OrderedUnitCostTaxInclusiveHome: any;
  // OrderedUnitCostTaxExclusiveForeign: any;
  // ForeignCurrency: any;
  // ForeignExchangeRate: any;
  // ForeignLandedUnitTaxExclusive: any;
  // TaxID: any;
  // TaxRate: any;
  // LineTotalTaxAmount: any;
  // UnitCostTaxInclusiveHome: any;
}